project-kural/requirements.txt
# Core Dependencies
gradio>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
scipy>=1.10.0

# LangChain Core and Community Packages (Updated for v0.2+)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3

# AI/ML Dependencies
openai-whisper>=20231117
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# API Dependencies
requests>=2.31.0
openai>=1.3.0

# Audio Processing
gTTS>=2.4.0
pydub>=0.25.1

# Testing Dependencies
pytest>=7.4.0
pytest-mock>=3.12.0

# Optional Dependencies for enhanced functionality
scikit-learn>=1.3.0
transformers>=4.35.0

# System Dependencies (install separately if needed)
# ffmpeg - required for whisper audio processing
# Install with: sudo apt-get install ffmpeg (Linux) or brew install ffmpeg (Mac)